swat_attacks/requirements.txt
amqp==5.2.0
asgiref==3.8.1
autopep8==2.1.0
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cramjam==2.8.3
Django==4.1.13
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.6.1
idna==3.7
kombu==5.3.7
prompt-toolkit==3.0.43
pycodestyle==2.11.1
pymongo==3.12.3
python-dateutil==2.9.0.post0
python-snappy==0.7.1
pytz==2024.1
redis==5.0.4
requests==2.32.0
six==1.16.0
sqlparse==0.2.4
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
